+++ components/Post/Post.jsx
@@ -5,16 +5,22 @@ import clsx from 'clsx';
 import Link from 'next/link';
 import { useMemo, useState } from 'react';
 import styles from './Post.module.css';
-import { CloseOutlined } from '@ant-design/icons';
+import { CheckCircleOutlined, CloseOutlined, EditOutlined } from '@ant-design/icons';
 import { useCallback } from 'react';
 import { usePostPages } from '@/lib/post';
 import { fetcher } from '@/lib/fetch';
 import toast from 'react-hot-toast';
-import { Button, Spin } from 'antd';
+import { Button, Spin, Typography } from 'antd';
+import PosterInner from '@/page-components/Post/PosterInner';
+import { useRouter } from 'next/router'
 
-const Post = ({ post, className, isDelete, isPublished }) => {
+const { Paragraph } = Typography;
+
+const Post = ({ post, className, isDelete = false, isPublished = false, isEdit = false }) => {
   const [isLoading, setIsLoading] = useState(false);
+  const [editMode, setEditMode] = useState(false);
   const { mutate } = usePostPages();
+  const router = useRouter();
 
   const timestampTxt = useMemo(() => {
     const diff = Date.now() - new Date(post.createdAt).getTime();
@@ -27,12 +33,18 @@ const Post = ({ post, className, isDelete, isPublished }) => {
       e.preventDefault();
       try {
         setIsLoading(true);
-        await fetcher('/api/posts', {
+        const res = await fetcher('/api/posts', {
           method: 'DELETE',
           headers: { 'Content-Type': 'application/json' },
           body: JSON.stringify({ id: post._id }),
         });
-        toast.success('You have deleted successfully');
+
+        let msg = 'You have deleted successfully.';
+        if (res?.comment?.deletedCount > 0) {
+          msg += ` Successfully deleted ${res?.comment?.deletedCount} ${res?.comment?.deletedCount <= 1 ? 'comment' : 'comments'} related to this post.`;
+        }
+        toast.success(msg, { duration: 4000 });
+
         // refresh post lists
         mutate();
       } catch (e) {
@@ -63,10 +75,18 @@ const Post = ({ post, className, isDelete, isPublished }) => {
     }
   }
 
+  const handleEdit = () => {
+    setEditMode(!editMode);
+  }
+
+  const handleSave = () => {
+    router.reload();
+  }
+
   return (
     <Spin spinning={isLoading} tip="Loading...">
       <div className={clsx(styles.root, className)}>
-        <div className={styles.creatorContainer}>
+        <div className={styles.header}>
           <Link href={`/user/${post.creator.username}`}>
             <div>
               <Container className={styles.creator}>
@@ -83,14 +103,36 @@ const Post = ({ post, className, isDelete, isPublished }) => {
             </div>
           </Link>
         </div>
-        <div className={styles.wrap}>
-          <p className={styles.content} title={post.content}>{post.content}</p>
-        </div>
+
+        {
+          editMode ? (
+            <PosterInner post={post} save={handleSave} cancel={() => setEditMode(false)} />
+          ) : (
+            <>
+              <Paragraph className={styles.title} ellipsis={!isEdit}>
+                {post.published && (
+                  <CheckCircleOutlined style={{ color: 'blue', marginRight: 8 }} />
+                )}
+                {post.title}
+              </Paragraph>
+              <div className={styles.wrap}>
+                <Paragraph ellipsis={!isEdit} className={styles.content}>{post.content}</Paragraph>
+              </div>
+            </>
+          )
+        }
+
         <div className={styles.wrap}>
           <time dateTime={String(post.createdAt)} className={styles.timestamp}>
             {timestampTxt}
           </time>
         </div>
+
+        {
+          isEdit && (
+            <Button type="text" shape='circle' icon={<EditOutlined />} className={styles.closeBtn} onClick={handleEdit} />
+          )
+        }
 
         {
           isDelete && (