+++ constants/apiPath.js
@@ -1,0 +1,3 @@
+export const API_POST = '/api/post';
+
+export default {};