+++ constants/index.js
@@ -1,0 +1,7 @@
+export const EMPTY_FUNCTION = () => { };
+
+export const PAGE_SIZE = 10;
+
+export const FIRST_PAGE = 1;
+
+export default {};