+++ constants/msg.js
@@ -1,0 +1,6 @@
+export const ERROR_MSG = {
+  USERNAME_OR_PASSWORD_IS_INCORRECT: 'Username or password is incorrect',
+  DATA_NOT_FOUND: 'Data not found',
+};
+
+export default {};