+++ constants/permission.js
@@ -1,0 +1,10 @@
+export const PERMISSION = {
+  POST: {
+    POST_VIEW: 'POST_VIEW',
+    POST_CREATE: 'POST_CREATE',
+    POST_EDIT: 'POST_EDIT',
+    POST_DELETE: 'POST_DELETE',
+  }
+};
+
+export default {};