+++ constants/routerPath.js
@@ -1,0 +1,8 @@
+export const ROUTER_PATH = {
+  HOME: '/',
+  LOGIN: '/login',
+  POST: '/post',
+  BLOG: '/blog',
+};
+
+export default {};