+++ page-components/Auth/Login.jsx
@@ -21,7 +21,7 @@ const Login = () => {
   const router = useRouter();
   useEffect(() => {
     if (isValidating) return;
-    if (user) router.replace('/feed');
+    if (user) router.replace('/');
   }, [user, router, isValidating]);
 
   const onSubmit = useCallback(