+++ page-components/Auth/SignUp.jsx
@@ -39,7 +39,7 @@ const SignUp = () => {
         });
         mutate({ user: response.user }, false);
         toast.success('Your account has been created');
-        router.replace('/feed');
+        router.replace('/');
       } catch (e) {
         toast.error(e.message);
       } finally {