+++ page-components/Post/PosterInner.jsx
@@ -1,4 +1,4 @@
-import { Button, Form, Input } from 'antd';
+import { Button, Form, Input, Space } from 'antd';
 import toast from 'react-hot-toast';
 import { useState, useCallback } from 'react';
 import { usePostPages } from '@/lib/post';
@@ -6,9 +6,10 @@ import { fetcher } from '@/lib/fetch';
 
 const { TextArea } = Input;
 
-const PosterInner = ({ user }) => {
+const PosterInner = ({ user = {}, post = {}, save = () => { }, cancel = () => { } }) => {
   const [isLoading, setIsLoading] = useState(false);
   const [form] = Form.useForm();
+  const editMode = post._id;
 
   const { mutate } = usePostPages();
 
@@ -16,15 +17,18 @@ const PosterInner = ({ user }) => {
     async (values) => {
       try {
         setIsLoading(true);
+        const requestBody = editMode ? { ...values, id: post._id } : values;
+        const msg = editMode ? 'You have saved successfully' : 'You have posted successfully';
         await fetcher('/api/posts', {
-          method: 'POST',
+          method: editMode ? 'PUT' : 'POST',
           headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(values),
+          body: JSON.stringify(requestBody),
         });
-        toast.success('You have posted successfully');
+        toast.success(msg);
         form.resetFields();
         // refresh post lists
         mutate();
+        save();
       } catch (e) {
         toast.error(e.message);
       } finally {
@@ -51,6 +55,7 @@ const PosterInner = ({ user }) => {
       <Form.Item
         name="title"
         rules={[{ required: true, message: 'Please input post\'s title!' }]}
+        initialValue={post.title}
       >
         <Input placeholder={`What's on your post title, ${user.name}?`} />
       </Form.Item>
@@ -58,14 +63,24 @@ const PosterInner = ({ user }) => {
       <Form.Item
         name="content"
         rules={[{ required: true, message: 'Please input post\'s content!' }]}
+        initialValue={post.content}
       >
         <TextArea rows={5} placeholder={`What's on your post content, ${user.name}?`} />
       </Form.Item>
 
       <Form.Item style={{ textAlign: 'center', margin: 0 }}>
-        <Button type="primary" htmlType="submit" shape="round" loading={isLoading}>
-          Submit
-        </Button>
+        <Space>
+          <Button type="primary" htmlType="submit" shape="round" loading={isLoading}>
+            {editMode ? 'Save' : 'Submit'}
+          </Button>
+          {
+            editMode && (
+              <Button shape="round" loading={isLoading} onClick={cancel}>
+                Cancel
+              </Button>
+            )
+          }
+        </Space>
       </Form.Item>
     </Form>
   )