+++ pages/api/posts/index.js
@@ -13,7 +13,8 @@ handler.get(async (req, res) => {
     req.db,
     req.query.before ? new Date(req.query.before) : undefined,
     req.query.by,
-    req.query.limit ? parseInt(req.query.limit, 10) : undefined
+    req.query.limit ? parseInt(req.query.limit, 10) : undefined,
+    req.query.published,
   );
 
   res.json({ posts });