+++ utils/index.js
@@ -1,0 +1,5 @@
+export const checkPermission = (data, role) => {
+  return !!data?.user?.roles?.[role] || !!data?.roles?.[role];
+}
+
+export default {};